refactor(controller): 컨트롤러 코어 및 프로세스 분리
